+++ sf-lifecycle/src/main/java/io/sunflower/lifecycle/Managed.java
@@ -3,6 +3,7 @@ package io.sunflower.lifecycle;
 /**
  * An interface for objects which need to be started and stopped as the application is started or
  * stopped.
+ * @author michael
  */
 public interface Managed {
 
@@ -11,12 +12,12 @@ public interface Managed {
    *
    * @throws Exception if something goes wrong; this will halt the application startup.
    */
-  void start() throws Exception;
+  default void start() throws Exception {}
 
   /**
    * Stops the object. Called <i>after</i> the application is no longer accepting requests.
    *
    * @throws Exception if something goes wrong.
    */
-  void stop() throws Exception;
+  default void stop() throws Exception {}
 }